Update bot.js
When the bot starts, a warning message now shows if lockdown mode is enabled.
A reason is now defined if a user is kicked for account being too young. (#10)
If a user is removed from the server while lockdown mode is enabled, the user will now receive a link in the removal message with more information.